Remove logger.py
